增加点赞

+++ src/main/java/com/project/miu/bean/model/OrderDetail.java
@@ -19,6 +19,6 @@ public class OrderDetail {
     private long uuid;
     private Timestamp createTime;
     private Timestamp updateTime;
-    private Byte delete;
+    private Integer delete;
 
 }
